+++ src/pbl.c
@@ -11,8 +11,8 @@
 
 // ---- Functions Definitions -----------------------------------------------------------------------------------------
 
-void PblAbortWithCriticalError(int exit_code, const char *string, const int line, const char* file_name) {
-  fprintf(stderr, "Traceback: \n File \"%s\", line %i", file_name, line);
+void PblThrowCriticalError(int exit_code, const char *string, const int line, const char* file_name) {
+  fprintf(stderr, "Traceback: \n File \"%s\", line %i \n", file_name, line);
   fprintf(stderr, "CRITICAL ERR [Para]: %s", string);
   exit(exit_code);
 }